requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==0.27.8
python-multipart==0.0.6